+++ R/utilities.R
@@ -51,11 +51,16 @@ fc_palette <- function(fc) {
 }
 
 update_n <- function(x, showCategory) {
+    if (!is.numeric(showCategory)) {
+        return(showCategory)
+    }
+
+    ## geneSets <- geneInCategory(x) ## use core gene for gsea result
     n <- showCategory
-    geneSets <- geneInCategory(x) ## use core gene for gsea result
     if (nrow(x) < n) {
         n <- nrow(x)
     }
+
     return(n)
 }
 
@@ -63,9 +68,11 @@ extract_geneSets <- function(x, n) {
     n <- update_n(x, n)
     geneSets <- geneInCategory(x) ## use core gene for gsea result
     y <- as.data.frame(x)
-    y <- y[1:n,]
     geneSets <- geneSets[y$ID]
     names(geneSets) <- y$Description
-    return(geneSets)
+    if (is.numeric(n)) {
+        return(geneSets[1:n])
+    }
+    return(geneSets[n]) ## if n is a vector of Description
 }